+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - added store pickup to shipping_service_list
 ### Changed
 - use constant instead of string as key for LA POST shipping service element when adding it to shipping_service_list
+- set plugin to use version 2.9.71 of cart-integration-sdk instead of 2.9.69
 
 ## 2.9.83 - 2017-10-27
 ### Changed